Split `ConfigureCompilerFlags.cmake`
Split `ConfigureCompilerFlags.cmake` into compiler and configuration
specific files, and set `CMAKE_Fortran_FLAGS_INIT` and
`CMAKE_Fortran_FLAGS_<CONFIG>_INIT` instead of `CMAKE_Fortran_FLAGS`
and `CMAKE_Fortran_FLAGS_<CONFIG>`. The new files are meant to be used
as toolchain files. This is a change of philosophy: we do not force
the choice of compiler options from the project, we leave them to the
user at configuration.