Do not try to find netCDF config
Do not try to find netCDF config file from the find module. This was
problematic in the case there is one version of NetCDF installed with
CMake (with the config file) and another without. It was then
difficult to choose the version installed without CMake. Also finding
the config file does not seem to be in the philosophy of the find
module.