Do not find NetCDF
Revert commit 4a6266e9. Change of philosophy. This was not
consistent. If NetCDF-Fortran were responsible for finding NetCDF then
NetCDF would be responsible for finding HDF5. Then we would have to
include a find HDF5 module. And so on. We have to rely on the
information given by pkg-config to set the correct compilation and
linking flags.