Set `PKG_CONFIG_ALLOW_SYSTEM_CFLAGS`
Set `PKG_CONFIG_ALLOW_SYSTEM_CFLAGS` instead of adding
`${netcdff_INCLUDEDIR}` to the include directories. I have checked
that this solution works, using the program `Test_NetCDF90`, on a PC
with Linux Mint, and on Ciclad, jean-zay and irene, with both ifort
and gfortran. Note that we save and restore the previous value of
`PKG_CONFIG_ALLOW_SYSTEM_CFLAGS`.